esppac: publish NAN when outside temperature is out of range
There are split AC units where the indoor unit supplies power to
the outdoor unit. When the AC is turned off, the indoor unit cuts
mains power to the outdoor unit. Since the outdoor unit can no longer
report anything, the indoor unit places an out-of-range temperature
value in place of the outdoor temperature, signalling that the
outdoor temperature is unavailable.